video_generation_tool/requirements.txt
google-generativeai
google-genai
moviepy
edge-tts
Pillow
python-dotenv
google-auth-oauthlib
google-api-python-client
